the brand new branch :smile:

+++ src/main/java/com/cutety/entity/Bill.java
@@ -4,6 +4,7 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -18,7 +19,7 @@ public class Bill {
 
     private String remark;
 
-    private Long consumptionAmount;
+    private BigDecimal consumptionAmount;
 
     private Integer userId;
